[dagster-airbyte] Scaffold DagsterAirbyteTranslator for rework
